refactor(kolme): extract block fallback reconciler module ownership (#1948)
